Clipboard functionality moved to composable
