+++ resources/js/Components/MapCardLineSmall.vue
@@ -1,6 +1,9 @@
 <script setup>
     import { computed, ref } from 'vue';
     import { Link } from '@inertiajs/vue3';
+    import { useClipboard } from '@/Composables/useClipboard';
+    
+    const { copy, copyState } = useClipboard();
 
     const props = defineProps({
         map: Object,
@@ -9,25 +12,6 @@
             default: true
         }
     });
-
-    const copyText = ref(false);
-
-
-    const copyMap = () => {
-        const textarea = document.createElement('textarea');
-        textarea.value = props.map?.name;
-        document.body.appendChild(textarea);
-        textarea.select();
-        document.execCommand('copy');
-        document.body.removeChild(textarea);
-
-        copyText.value = true;
-
-        setTimeout(() => {
-            copyText.value = false;
-        }, 1000);
-    };
-
 
     let weaponsList = props.map?.weapons?.split(',') ?? [];
 
@@ -93,7 +77,7 @@
                     <Link class="text-lg text-blue-400 hover:text-blue-300 font-bold" :href="route('maps.map', getRouteData)"> {{ mapName }} </Link>
                 
                     <Popper :closeDelay="300" hover style="z-index: 100;" >
-                        <div class="transition-all duration-500 ease-in-out opacity-0 group-hover:opacity-100 cursor-pointer text-gray-400 hover:text-green-500 ml-2" @click="copyMap">
+                        <div class="transition-all duration-500 ease-in-out opacity-0 group-hover:opacity-100 cursor-pointer text-gray-400 hover:text-green-500 ml-2" @click="copy(map?.name)">
                             <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-4 h-4">
                                 <path stroke-linecap="round" stroke-linejoin="round" d="M16.5 8.25V6a2.25 2.25 0 0 0-2.25-2.25H6A2.25 2.25 0 0 0 3.75 6v8.25A2.25 2.25 0 0 0 6 16.5h2.25m8.25-8.25H18a2.25 2.25 0 0 1 2.25 2.25V18A2.25 2.25 0 0 1 18 20.25h-7.5A2.25 2.25 0 0 1 8.25 18v-1.5m8.25-8.25h-6a2.25 2.25 0 0 0-2.25 2.25v6" />
                             </svg>                      
@@ -103,7 +87,7 @@
                         <template #content>
                             <div class="px-4 py-2 rounded-md bg-blackop-80">
                                 <div class="text-gray-400">
-                                    <span v-if="!copyText">Copy</span>
+                                    <span v-if="!copyState">Copy</span>
                                     <span v-else class="text-green-400">Copied !</span>
                                 </div>
                             </div>
