+++ resources/js/Components/MapDetails.vue
@@ -1,28 +1,14 @@
 <script setup>
     import { computed, ref, watchEffect } from 'vue';
     import { Link } from '@inertiajs/vue3';
+    import { useClipboard } from '@/Composables/useClipboard';
+    
+    const { copy, copyState } = useClipboard();
 
     const props = defineProps({
         map: Object,
         mapname: String
     });
-
-    const copyText = ref(false);
-
-    const copyMap = () => {
-        const textarea = document.createElement('textarea');
-        textarea.value = props.map?.name ?? props.mapname;
-        document.body.appendChild(textarea);
-        textarea.select();
-        document.execCommand('copy');
-        document.body.removeChild(textarea);
-
-        copyText.value = true;
-
-        setTimeout(() => {
-            copyText.value = false;
-        }, 1000);
-    };
 
     let weaponsList = ref([]);
     let itemsList = ref([]);
@@ -101,7 +87,7 @@
                 <Link :class="(mapName.length > 30) ? 'text-sm hover:text-blue-300 font-bold mh-50' : 'normal-map-name-text hover:text-blue-300 font-bold mh-50'" :href="route('maps.map', getRouteData)"> {{ mapName }} </Link>
             
                 <Popper :closeDelay="300" hover style="z-index: 3;" >
-                    <div class="transition-all duration-500 ease-in-out opacity-0 group-hover:opacity-100 cursor-pointer text-gray-400 hover:text-green-500 ml-2" @click="copyMap">
+                    <div class="transition-all duration-500 ease-in-out opacity-0 group-hover:opacity-100 cursor-pointer text-gray-400 hover:text-green-500 ml-2" @click="copy(map?.name ?? mapname)">
                         <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-4 h-4">
                             <path stroke-linecap="round" stroke-linejoin="round" d="M16.5 8.25V6a2.25 2.25 0 0 0-2.25-2.25H6A2.25 2.25 0 0 0 3.75 6v8.25A2.25 2.25 0 0 0 6 16.5h2.25m8.25-8.25H18a2.25 2.25 0 0 1 2.25 2.25V18A2.25 2.25 0 0 1 18 20.25h-7.5A2.25 2.25 0 0 1 8.25 18v-1.5m8.25-8.25h-6a2.25 2.25 0 0 0-2.25 2.25v6" />
                         </svg>                      
@@ -111,7 +97,7 @@
                     <template #content>
                         <div class="copy-text px-4 py-2 rounded-md bg-blackop-80">
                             <div class="text-gray-400">
-                                <span v-if="!copyText">Copy</span>
+                                <span v-if="!copyState">Copy</span>
                                 <span v-else class="text-green-400">Copied!</span>
                             </div>
                         </div>
