+++ resources/js/Components/MapSearchItem.vue
@@ -1,5 +1,8 @@
 <script setup>
     import { router } from '@inertiajs/vue3';
+    import { useClipboard } from '@/Composables/useClipboard';
+    
+    const { copy } = useClipboard();
 
     const props = defineProps({
         map: Object,
@@ -9,16 +12,6 @@
         },
         click: Function
     });
-    
-    const copyMap = () => {
-        const textarea = document.createElement('textarea');
-        textarea.value = props.map?.name ?? props.mapname;
-        document.body.appendChild(textarea);
-        textarea.select();
-        document.execCommand('copy');
-        document.body.removeChild(textarea);
-    };
-
 
     let weaponsList = props.map?.weapons?.split(',') ?? [];
 
@@ -67,7 +60,7 @@
                     <div class="flex items-center">
                         <div class="text-md text-blue-400 font-bold"> {{ map.name }} </div>
                     
-                        <div class="transition-all duration-500 ease-in-out opacity-0 group-hover:opacity-100 cursor-pointer text-gray-400 hover:text-green-500 ml-2" @click="copyMap">
+                        <div class="transition-all duration-500 ease-in-out opacity-0 group-hover:opacity-100 cursor-pointer text-gray-400 hover:text-green-500 ml-2" @click="copy(map?.name ?? mapname)">
                             <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-4 h-4">
                                 <path stroke-linecap="round" stroke-linejoin="round" d="M16.5 8.25V6a2.25 2.25 0 0 0-2.25-2.25H6A2.25 2.25 0 0 0 3.75 6v8.25A2.25 2.25 0 0 0 6 16.5h2.25m8.25-8.25H18a2.25 2.25 0 0 1 2.25 2.25V18A2.25 2.25 0 0 1 18 20.25h-7.5A2.25 2.25 0 0 1 8.25 18v-1.5m8.25-8.25h-6a2.25 2.25 0 0 0-2.25 2.25v6" />
                             </svg>
