+++ resources/js/Pages/Profile.vue
@@ -3,6 +3,9 @@
     import { Head, router, Link } from '@inertiajs/vue3';
     import ProfileRecord from '@/Components/ProfileRecord.vue';
     import Pagination from '@/Components/Basic/Pagination.vue';
+    import { useClipboard } from '@/Composables/useClipboard';
+    
+    const { copy, copyState } = useClipboard();
 
     const props = defineProps({
         user: Object,
@@ -166,24 +169,6 @@
         }
     })
 
-    const copyText = ref(false);
-
-
-    const copyHandle = (text) => {
-        const textarea = document.createElement('textarea');
-        textarea.value = text
-        document.body.appendChild(textarea);
-        textarea.select();
-        document.execCommand('copy');
-        document.body.removeChild(textarea);
-
-        copyText.value = true;
-
-        setTimeout(() => {
-            copyText.value = false;
-        }, 1000);
-    };
-
 </script>
 
 <template>
@@ -238,11 +223,11 @@
                                 <span>TWITCH</span>
                             </a>
 
-                            <div @click="copyHandle(user?.discord_name)" v-if="user?.discord_name" class="flex items-center gap-2 cursor-pointer text-xs font-medium me-2 px-2.5 py-1.5 rounded bg-indigo-900 text-indigo-300 hover:text-indigo-100">
+                            <div @click="copy(user?.discord_name)" v-if="user?.discord_name" class="flex items-center gap-2 cursor-pointer text-xs font-medium me-2 px-2.5 py-1.5 rounded bg-indigo-900 text-indigo-300 hover:text-indigo-100">
                                 <svg class="w-5 h-5" width="800px" height="800px" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
                                     <path d="M18.59 5.88997C17.36 5.31997 16.05 4.89997 14.67 4.65997C14.5 4.95997 14.3 5.36997 14.17 5.69997C12.71 5.47997 11.26 5.47997 9.83001 5.69997C9.69001 5.36997 9.49001 4.95997 9.32001 4.65997C7.94001 4.89997 6.63001 5.31997 5.40001 5.88997C2.92001 9.62997 2.25001 13.28 2.58001 16.87C4.23001 18.1 5.82001 18.84 7.39001 19.33C7.78001 18.8 8.12001 18.23 8.42001 17.64C7.85001 17.43 7.31001 17.16 6.80001 16.85C6.94001 16.75 7.07001 16.64 7.20001 16.54C10.33 18 13.72 18 16.81 16.54C16.94 16.65 17.07 16.75 17.21 16.85C16.7 17.16 16.15 17.42 15.59 17.64C15.89 18.23 16.23 18.8 16.62 19.33C18.19 18.84 19.79 18.1 21.43 16.87C21.82 12.7 20.76 9.08997 18.61 5.88997H18.59ZM8.84001 14.67C7.90001 14.67 7.13001 13.8 7.13001 12.73C7.13001 11.66 7.88001 10.79 8.84001 10.79C9.80001 10.79 10.56 11.66 10.55 12.73C10.55 13.79 9.80001 14.67 8.84001 14.67ZM15.15 14.67C14.21 14.67 13.44 13.8 13.44 12.73C13.44 11.66 14.19 10.79 15.15 10.79C16.11 10.79 16.87 11.66 16.86 12.73C16.86 13.79 16.11 14.67 15.15 14.67Z" fill="#ffffff"/>
                                 </svg>
-                                <span class="text-sm" v-if="!copyText">#{{user?.discord_name}}</span>
+                                <span class="text-sm" v-if="!copyState">#{{user?.discord_name}}</span>
                                 <span v-else class="text-white font-bold">COPIED</span>
                             </div>
